changes to setup.py and monosat_c.py to support Linux, Windows, and OSX

+++ api/python/setup.py
@@ -1,7 +1,9 @@
 #!/usr/bin/env python3
-
+from __future__ import print_function
 from distutils.core import setup
 
+
+import sys
 import os.path
 import shutil
 import platform
@@ -9,10 +11,16 @@ import platform
 if platform.system != "Windows":
     copy_lib = "monosat/libmonosat.so"
     orig_lib = "../../SharedLibrary/libmonosat.so"
-    if not os.path.exists(copy_lib):
-        shutil.copy2(orig_lib, "monosat")
-    if os.path.getmtime(orig_lib) > os.path.getmtime(copy_lib):
-        shutil.copy2(orig_lib, "monosat")
+    if not os.path.exists(orig_lib) and os.path.exists("../../OSX_SharedLibrary/libmonosat.so"):
+        orig_lib = "../../OSX_SharedLibrary/libmonosat.so"
+    if os.path.exists(orig_lib):                     
+        if not os.path.exists(copy_lib):
+            shutil.copy2(orig_lib, "monosat")
+        if  os.path.getmtime(orig_lib) > os.path.getmtime(copy_lib):
+            shutil.copy2(orig_lib, "monosat")
+
+if not os.path.exists(copy_lib):
+    print("Warning: could not find libmonosat.so or libmonosat.dll. See README for instructions on compiling the library, the re-install",file=sys.stderr)
 
 setup(name='monosat',
       version='1.0',
